Fixed bug in ActionQueue, Changes to threads and queued ftp actions
There was a bug in ActionQueue.clear() method that locked the db preventing
its use. Reorganized the threading code and added startCheckout timer for
FtpObject and  FileWatcher classes. Lastly, added queues in the FtpObject to
ultimately make sure deleteFile, uploadFile and downloadFile are called sequentially
in the same thread, to prevent FTP or FTP_TLS from raising erros