Enhanced sync functionality
Problem: After download, the local file will always
have a mod date newer.
I changed the downloadFile a bit so the
database is updated quickly with the new date
for both locamdate and servermdate, this way the change
less noticeable after a download. Added a tolerance of 1 sec
to the onChanged slot.
Added a fallback sync on SyncCore which is trigger when
the FtpObject is done.